improvement: pre-compilation with vite mode option
closes #29
